+++ src/bitmap/pcfwrite.c
@@ -236,7 +236,8 @@ pcfWriteFont(FontPtr pFont, FontFilePtr file)
     }
     offsetProps = malloc(pFont->info.nprops * sizeof(FontPropRec));
     if (!offsetProps) {
-      pcfError("pcfWriteFont(): Couldn't allocate offsetProps (%d*%d)", pFont->info.nprops, sizeof(FontPropRec));
+	pcfError("pcfWriteFont(): Couldn't allocate offsetProps (%d*%d)",
+		 pFont->info.nprops, (int) sizeof(FontPropRec));
 	return AllocError;
     }
     prop_string_size = 0;
